+++ portal/src/CheckboxWithTooltip.module.scss
@@ -3,9 +3,5 @@
 }
 
 .tooltip {
-  max-width: 300px;
-}
-
-.infoIcon {
-  padding: 0 5px 5px;
+  display: flex;
 }
